+++ src/Common/Util.cpp
@@ -191,3 +191,47 @@ std::uint32_t dcclite::ntohl(const std::uint32_t v) noexcept
 {
 	return ::ntohl(v);
 }
+
+#ifdef WIN32
+//https://learn.microsoft.com/pt-br/previous-versions/visualstudio/visual-studio-2015/debugger/how-to-set-a-thread-name-in-native-code?view=vs-2015&redirectedfrom=MSDN
+//
+//  
+// Usage: SetThreadName ((DWORD)-1, "MainThread");  
+//  
+
+const DWORD MS_VC_EXCEPTION = 0x406D1388;
+#pragma pack(push,8)  
+typedef struct tagTHREADNAME_INFO
+{
+	DWORD dwType; // Must be 0x1000.  
+	LPCSTR szName; // Pointer to name (in user addr space).  
+	DWORD dwThreadID; // Thread ID (-1=caller thread).  
+	DWORD dwFlags; // Reserved for future use, must be zero.  
+} THREADNAME_INFO;
+#pragma pack(pop)  
+void dcclite::SetThreadName(std::thread &thread, const char *threadName) 
+{
+	HANDLE h = thread.native_handle();
+
+	THREADNAME_INFO info;
+	info.dwType = 0x1000;
+	info.szName = threadName;
+	info.dwThreadID = GetThreadId(h);
+	info.dwFlags = 0;
+#pragma warning(push)  
+#pragma warning(disable: 6320 6322)  
+	__try {
+		RaiseException(MS_VC_EXCEPTION, 0, sizeof(info) / sizeof(ULONG_PTR), (ULONG_PTR *)&info);
+	}
+	__except (EXCEPTION_EXECUTE_HANDLER) {
+	}
+#pragma warning(pop)  
+}
+
+#else
+void dcclite::SetThreadName(std::thread &thread, const char *threadName)
+{
+	//not implemented
+}
+
+#endif